Refactoring: reorganize Simulation code, split it into utils, engine, extract types, etc.
[#171860513]
