+++ src/components/terrain-panel.tsx
@@ -6,7 +6,7 @@ import { renderZones } from "./zone-selector";
 import { TerrainTypeSelector } from "./terrain-type-selector";
 import { VegetationSelector } from "./vegetation-selector";
 import { DroughtSelector } from "./drought-selector";
-import { TerrainType, Vegetation } from "../models/fire-model";
+import { TerrainType, Vegetation } from "../types";
 import { WindCircularControl } from "./wind-circular-control";
 import { TerrainSummary } from "./terrain-summary";
