+++ src/components/terrain-type-selector.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Radio, RadioGroup, FormControlLabel } from "@material-ui/core";
-import { TerrainType } from "../models/fire-model";
+import { TerrainType } from "../types";
 
 import * as css from "./terrain-type-selector.scss";
