+++ src/models/engine/get-fire-spread-rate.ts
@@ -1,33 +1,8 @@
-import { Fuel } from "../types";
+import { Fuel } from "../../types";
 import { Vector2 } from "three";
+import { Vegetation, IWindProps } from "../../types";
 
-export enum Vegetation {
-  Grass = 0,
-  Shrub = 1,
-  ForestSmallLitter = 2,
-  ForestLargeLitter = 3
-}
-export enum TerrainType {
-  Plains = 0,
-  Foothills = 1,
-  Mountains = 2
-}
-export enum DroughtLevel {
-  NoDrought = 0,
-  MildDrought = 1,
-  MediumDrought = 2,
-  SevereDrought = 3
-}
-
-export interface IWindProps {
-  // Wind speed in mph.
-  speed: number;
-  // Angle in degrees following this definition: https://en.wikipedia.org/wiki/Wind_direction
-  // 0 is northern wind, 90 is eastern wind.
-  direction: number;
-}
-
-export interface ICellProps {
+interface ICellProps {
   x: number;
   y: number;
   vegetation: Vegetation;
@@ -71,14 +46,6 @@ const FuelConstants: {[key in Vegetation]: Fuel} = {
     packingRatio: 0.02224,
     mx: 0.25
   }
-};
-
-// values for each level of vegetation: Grass, Shrub, ForestSmallLitter, ForestLargeLitter
-export const moistureLookups: {[key in DroughtLevel]: number[]} = {
-  [DroughtLevel.NoDrought]: [0.1275, 0.255, 0.17, 0.2125],
-  [DroughtLevel.MildDrought]: [0.09, 0.18, 0.12, 0.15],
-  [DroughtLevel.MediumDrought]: [0.0525, 0.105, 0.07, 0.0875],
-  [DroughtLevel.SevereDrought]: [0.015, 0.03, 0.02, 0.025],
 };
 
 // Helper vector used repeatedly in other calculations.
