+++ src/models/zone.ts
@@ -1,4 +1,4 @@
-import { Vegetation, TerrainType, DroughtLevel, moistureLookups } from "./fire-model";
+import { Vegetation, TerrainType, DroughtLevel } from "../types";
 import { observable } from "mobx";
 
 export interface ZoneOptions {
@@ -6,6 +6,15 @@ export interface ZoneOptions {
   terrainType?: TerrainType;
   droughtLevel?: number;
 }
+
+// values for each level of vegetation: Grass, Shrub, ForestSmallLitter, ForestLargeLitter
+const moistureLookups: {[key in DroughtLevel]: number[]} = {
+  [DroughtLevel.NoDrought]: [0.1275, 0.255, 0.17, 0.2125],
+  [DroughtLevel.MildDrought]: [0.09, 0.18, 0.12, 0.15],
+  [DroughtLevel.MediumDrought]: [0.0525, 0.105, 0.07, 0.0875],
+  [DroughtLevel.SevereDrought]: [0.015, 0.03, 0.02, 0.025],
+};
+
 
 export class Zone {
   @observable public vegetation: Vegetation = Vegetation.Grass;
