+++ src/types.ts
@@ -12,3 +12,31 @@ export interface Town {
   name: string;
   position: Vector2;
 }
+
+export enum Vegetation {
+  Grass = 0,
+  Shrub = 1,
+  ForestSmallLitter = 2,
+  ForestLargeLitter = 3
+}
+
+export enum TerrainType {
+  Plains = 0,
+  Foothills = 1,
+  Mountains = 2
+}
+
+export enum DroughtLevel {
+  NoDrought = 0,
+  MildDrought = 1,
+  MediumDrought = 2,
+  SevereDrought = 3
+}
+
+export interface IWindProps {
+  // Wind speed in mph.
+  speed: number;
+  // Angle in degrees following this definition: https://en.wikipedia.org/wiki/Wind_direction
+  // 0 is northern wind, 90 is eastern wind.
+  direction: number;
+}
